deteccao-conflitos/requirements.txt
faiss-cpu
openai
tiktoken
